card generation refactoring
